update structure of test files separate tests under appropriate test files
